Struct nearswap::NearCLP

pub struct NearCLP {
    pub fee_dst: AccountId,
    pub owner: AccountId,
    // some fields omitted
}

NearCLP is the main contract for managing the swap pools and liquidity. It implements the NEARswap functionality.

Fields
fee_dst: AccountIdowner: AccountId
Implementations

impl NearCLP

pub fn new(owner: ValidAccountId) -> Self

pub fn set_fee_dst(&mut self, fee_dst: ValidAccountId)

Updates the fee destination destination account

pub fn change_owner(&mut self, new_owner: ValidAccountId)

Owner is an account (can be a multisig) who has management rights to update fee size.

pub fn create_pool(&mut self, token: ValidAccountId)

Allows any user to creat a new near-token pool. Each pool is identified by the token account - which we call the Pool Reserve Token. If a pool for give token exists then "E1" assert exception is thrown. TODO: charge user for a storage created!

pub fn pool_info(&self, token: &AccountId) -> Option<PoolInfo>

Extracts public information of the token pool.

pub fn list_pools(&self) -> Vec<AccountId>

Returns list of pools identified as their reserve token AccountId.

pub fn add_liquidity(
&mut self,
token: AccountId,
max_tokens: U128,
min_shares: U128
) -> U128

Increases Near and the Reserve token liquidity. The supplied funds must preserve current ratio of the liquidity pool. Returns amount of LP Shares is minted for the user.

pub fn withdraw_liquidity(
&mut self,
token: AccountId,
shares: U128,
min_ynear: U128,
min_tokens: U128
)

Redeems shares for liquidity stored in this pool with condition of getting at least min_ynear of Near and min_tokens of reserve tokens (token). Shares are note exchengable between different pools.

pub fn swap_near_to_token_exact_in(
&mut self,
token: AccountId,
min_tokens: U128
)

Swaps NEAR to token and transfers the reserve tokens to the caller. Caller attaches near tokens he wants to swap to the transacion under a condition of receving at least min_tokens of token.

pub fn swap_near_to_token_exact_in_xfr(
&mut self,
token: AccountId,
min_tokens: U128,
recipient: AccountId
)

Same as swap_near_to_token_exact_in, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_near_to_token_exact_out(
&mut self,
token: AccountId,
tokens_out: U128
)

Swaps NEAR to token and transfers the reserve tokens to the caller. Caller attaches maximum amount of NEAR he is willing to swap to receive tokens_out of token wants to swap to the transacion. Surplus of NEAR tokens will be returned. Transaction will panic if the caller doesn't attach enough NEAR tokens.

pub fn swap_near_to_token_exact_out_xfr(
&mut self,
token: AccountId,
tokens_out: U128,
recipient: AccountId
)

Same as swap_near_to_token_exact_out, but user additionly specifies the recipient who will receive the reserve tokens after the swap.

pub fn swap_token_to_near_exact_in(
&mut self,
token: AccountId,
tokens_paid: U128,
min_ynear: U128
)

Swaps tokens_paid of token to NEAR and transfers NEAR to the caller under acc condition of receving at least min_ynear yocto NEARs. Preceeding to this transaction, caller has to create sufficient allowance of token for this contract (at least tokens_paid). TODO: Transaction will panic if a caller doesn't provide enough allowance.

pub fn swap_token_to_near_exact_in_xfr(
&mut self,
token: AccountId,
tokens_paid: U128,
min_ynear: U128,
recipient: AccountId
)

Same as swap_token_to_near_exact_in, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_token_to_near_exact_out(
&mut self,
token: AccountId,
ynear_out: U128,
max_tokens: U128
)

Swaps token to NEAR and transfers NEAR to the caller. Caller defines the amount of NEAR he wants to receive under a condition of not spending more than max_tokens of token. Preceeding to this transaction, caller has to create sufficient allowance of token for this contract. TODO: Transaction will panic if a caller doesn't provide enough allowance.

pub fn swap_token_to_near_exact_out_xfr(
&mut self,
token: AccountId,
ynear_out: U128,
max_tokens: U128,
recipient: AccountId
)

Same as swap_token_to_near_exact_out, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_tokens_exact_in(
&mut self,
from: AccountId,
to: AccountId,
tokens_in: U128,
min_tokens_out: U128
)

Swaps two different tokens. Caller defines the amount of tokens he wants to swap under a condition of receving at least min_to_tokens. Preceeding to this transaction, caller has to create a sufficient allowance of from token for this contract. Transaction will panic if a caller doesn't provide enough allowance.

pub fn swap_tokens_exact_in_xfr(
&mut self,
from: AccountId,
to: AccountId,
tokens_in: U128,
min_tokens_out: U128,
recipient: AccountId
)

Same as swap_tokens_exact_in, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn swap_tokens_exact_out(
&mut self,
from: AccountId,
to: AccountId,
tokens_out: U128,
max_tokens_in: U128
)

Swaps two different tokens. Caller defines the amount of tokens he wants to receive under a condiiton of not spending more than max_from_tokens. Preceeding to this transaction, caller has to create a sufficient allowance of from token for this contract. Transaction will panic if a caller doesn't provide enough allowance.

pub fn swap_tokens_exact_out_xfr(
&mut self,
from: AccountId,
to: AccountId,
tokens_out: U128,
max_tokens_in: U128,
recipient: AccountId
)

Same as swap_tokens_exact_out, but user additionly specifies the recipient who will receive the tokens after the swap.

pub fn price_near_to_token_in(&self, token: AccountId, ynear_in: U128) -> U128

Calculates amount of tokens user will recieve when swapping ynear_in for token assets

pub fn price_near_to_token_out(
&self,
token: AccountId,
tokens_out: U128
) -> U128

Calculates amount of NEAR user will need to swap if he wants to receive tokens_out of token

pub fn price_token_to_near_in(&self, token: AccountId, tokens_in: U128) -> U128

Calculates amount of NEAR user will recieve when swapping tokens_in for NEAR.

pub fn price_token_to_near_out(&self, token: AccountId, ynear_out: U128) -> U128

Calculates amount of tokens user will need to swap if he wants to receive tokens_out of tokens

pub fn price_token_to_token_in(
&self,
from: AccountId,
to: AccountId,
tokens_in: U128
) -> U128

Calculates amount of tokens to user will receive when swapping tokens_in of from

pub fn price_token_to_token_out(
&self,
from: AccountId,
to: AccountId,
tokens_out: U128
) -> U128

Calculates amount of tokens from user will need to swap if he wants to receive tokens_out of tokens to

pub fn add_liquidity_transfer_callback(&mut self, token: AccountId)

pub fn token_url(&self, token: AccountId) -> String

returns resource to more information about the token.

pub fn granularity(&self, token: AccountId) -> U128

granularity is the smallest amount of tokens (in the internal denomination) which may be minted, sent or burned at any time.

pub fn decimals(&self, token: AccountId) -> u8

Returns the number of decimals the token uses - e.g. 8, means to divide the token amount by 100000000 to get its user representation.

pub fn total_supply(&self, token: AccountId) -> U128

Returns total supply of given subtoken. Implements the NEP-MFT standard.

pub fn balance_of(&self, token: AccountId, holder: AccountId) -> U128

Returns the owner shares balance of a pool identified by the token.

pub fn transfer(
&mut self,
token: String,
recipient: AccountId,
amount: U128,
msg: String,
memo: String
) -> bool

Transfer amount of LP Shares (Liquidity Provider Shares) of a pool identified by the token (must be a valid AccountID) from the predecessor to the recipient account. Implements the NEP-MFT interface. If recipient is a smart-contract, then transfer_call should be used instead. recipient MUST NOT be a smart-contract. msg is a message for recipient. It might be used to send additional call memo: arbitrary data with no specified format used to link the transaction with an external data. If referencing a binary data, it should use base64 serialization. The function panics if the token doesn't refer to any registered pool or the predecessor doesn't have sufficient amount of shares.

pub fn transfer_call(
&mut self,
token: String,
recipient: AccountId,
amount: U128,
msg: String,
memo: String
) -> bool

Transfer amount of LP Shares (Liquidity Provider Shares) of a pool identified by the token (must be a valid AccountID) from the predecessor to the recipient contract. Implements the NEP-MFT interface. recipient MUST be a smart contract address. The recipient contract MUST implement MFTRecipient interface. msg is a message sent to the recipient. It might be used to send additional call memo: arbitrary data with no specified format used to link the transaction with an external event. If referencing a binary data, it should use base64 serialization. The function panics if the token doesn't refer to any registered pool or the predecessor doesn't have sufficient amount of shares.

pub fn remove_pool(&mut self, token: AccountId)