Module nearswap::util

Structs

u256
Little-endian large integer type

Constants

MAX_GAS
Prepaid gas costs. TODO: we need to adjust this value properly.

NDENOM
Near denomination = 1e24. Usage: { amount: 50*E24 }

NEP21_STORAGE_DEPOSIT
nep21 may require NEAR deposit for storage to create a new balance

STORAGE_BYTE_PRICE
Price per 1 byte of storage from mainnet genesis config. 100e18

TGAS
TGas denomination 1 Tera Gas => 1e12 yocto Nears

Functions

assert_account_is_valid
Ensures that an account a is valid and panics if it's not. name: printed name of the account

is_promise_success

yton
yoctoNEAR to NEAR. Rounds to nearest.